console: Move requireRequest to view
