+++ pkg/webui/console/components/device-import-form/fallback-version-ids-section.js
@@ -14,24 +14,15 @@
 
 import React from 'react'
 import { Col, Row } from 'react-grid-system'
-import { useSelector } from 'react-redux'
 
 import { useFormContext } from '@ttn-lw/components/form'
-import ErrorNotification from '@ttn-lw/components/error-notification'
-
-import RequireRequest from '@ttn-lw/lib/components/require-request'
 
 import OtherHint from '@console/containers/device-profile-section/hints/other-hint'
 import VersionIdsSection from '@console/containers/device-profile-section'
 
 import { hasSelectedDeviceRepositoryOther } from '@console/lib/device-utils'
 
-import { listBrands } from '@console/store/actions/device-repository'
-
-import { selectSelectedApplicationId } from '@console/store/selectors/applications'
-
 const FallbackVersionIdsSection = () => {
-  const appId = useSelector(selectSelectedApplicationId)
   const { values } = useFormContext()
   const { version_ids } = values
   const version = version_ids
@@ -41,14 +32,8 @@ const FallbackVersionIdsSection = () => {
   return (
     <Row>
       <Col>
-        <RequireRequest
-          requestAction={listBrands(appId, {}, ['name', 'lora_alliance_vendor_id'])}
-          errorRenderFunction={ErrorNotification}
-          spinnerProps={{ center: false, inline: true }}
-        >
-          <VersionIdsSection />
-          {showOtherHint && <OtherHint manualGuideDocsPath="/devices/adding-devices/" />}
-        </RequireRequest>
+        <VersionIdsSection />
+        {showOtherHint && <OtherHint manualGuideDocsPath="/devices/adding-devices/" />}
       </Col>
     </Row>
   )
